Improved the player health

+++ Tile.pde
@@ -112,7 +112,7 @@ class Tile extends BaseObject{
     breakSound.play();
   }
 
-  void setMaxHp(int hpToSet){
+  void setMaxHp(float hpToSet){
     maxHp = hpToSet;
     hp = hpToSet;
   }
